Updated dependencies and composer package definition
- Updated minimum supported PHP version to 5.6.
- Updated all components and modules to set the minimum supported version to the
  most recent stable release; for those with v3 releases, these are supported
  via `||` notation.
- Moved `Module` class into the `src/` tree, where it belongs.
  - Removed the `getAutoloaderConfig()` implementation, as it was redundant.
  - Fixed how it attaches the content validation listener aggregate to the event
    manager.
- Made PHPUnit configuration a dist file.
- Enabled the lowest/locked/latest CI test strategy.
