+++ .gitignore
@@ -1,2 +1,2 @@
-composer.lock
 vendor/
+phpunit.xml
