--- Module.php
@@ -1,39 +1,0 @@
-<?php
-/**
- * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2014 Zend Technologies USA Inc. (http://www.zend.com)
- */
-
-namespace ZF\ContentValidation;
-
-use Zend\Loader\StandardAutoloader;
-use Zend\Mvc\MvcEvent;
-use ZF\ContentValidation\ContentValidationListener;
-
-class Module
-{
-    public function getAutoloaderConfig()
-    {
-        return [
-            StandardAutoloader::class => [
-                'namespaces' => [
-                    __NAMESPACE__ => __DIR__ . '/src/',
-                ],
-            ],
-        ];
-    }
-
-    public function getConfig()
-    {
-        return include __DIR__ . '/config/module.config.php';
-    }
-
-    public function onBootstrap(MvcEvent $e)
-    {
-        $app      = $e->getApplication();
-        $events   = $app->getEventManager();
-        $services = $app->getServiceManager();
-
-        $events->attach($services->get(ContentValidationListener::class));
-    }
-}
