+++ src/Module.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
+ * @copyright Copyright (c) 2014-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ */
+
+namespace ZF\ContentValidation;
+
+use Zend\Mvc\MvcEvent;
+use ZF\ContentValidation\ContentValidationListener;
+
+class Module
+{
+    public function getConfig()
+    {
+        return include __DIR__ . '/../config/module.config.php';
+    }
+
+    public function onBootstrap(MvcEvent $e)
+    {
+        $app      = $e->getApplication();
+        $events   = $app->getEventManager();
+        $services = $app->getServiceManager();
+
+        $services->get(ContentValidationListener::class)->attach($events);
+    }
+}
